eval/pyproject.toml
[project]
name = "srm-eval"
version = "0.1.0"
description = "Distribution-level speech representation evaluation"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "scipy>=1.10",
    "torch>=2.4",
    "torchaudio>=2.4",
    "transformers>=4.40",
    "huggingface-hub>=0.20",
    "tqdm>=4.66",
    "typer>=0.12",
    "pyyaml>=6.0",
    "rich>=13.0",
]
dev = [
    "pytest>=8.0",
]

[project.scripts]
srm-eval = "srm_eval.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
